group buttons

+++ app/src/main/java/com/example/restuarant/extentions/Extentions.kt
@@ -63,7 +63,7 @@ typealias SingleBlock <T> = (T) -> Unit
 fun View.visible(visible: Boolean) {
     this.visibility = if (visible) View.VISIBLE else View.GONE
 }
-fun stringFormat(string: Int):String{
+fun stringFormat(string: Long):String{
     return String.format("%,d", string).replace(',', ' ')
 }
 fun Fragment.showSnackMessage(message: String) {
